rename stat as prom package
